docs: AGENTS.md 增加 Git 提交规则 - 代码修改后必须提交到 dev 分支

+++ navatation-business/src/main/java/com/navatation/business/dto/req/nav/SortItemDTO.java
@@ -1,6 +1,7 @@
 package com.navatation.business.dto.req.nav;
 
 import lombok.Data;
+import java.math.BigDecimal;
 import com.navatation.business.dto.req.nav.SortItemDTO;
 
 /**
@@ -11,5 +12,5 @@ import com.navatation.business.dto.req.nav.SortItemDTO;
 @Data
 public class SortItemDTO {
     private String shortcutId;
-    private Double sortOrder;
+    private BigDecimal sortOrder;
 }
