+++ navatation-business/src/main/java/com/navatation/business/dto/resp/recommend/RecommendCategoryRespDTO.java
@@ -1,6 +1,7 @@
 package com.navatation.business.dto.resp.recommend;
 
 import lombok.Data;
+import java.math.BigDecimal;
 import java.util.List;
 import com.navatation.business.dto.resp.recommend.RecommendSiteRespDTO;
 import com.navatation.business.dto.resp.recommend.RecommendCategoryRespDTO;
@@ -15,6 +16,6 @@ public class RecommendCategoryRespDTO {
     private String categoryId;
     private String categoryName;
     private String categoryIcon;
-    private Double sortOrder;
+    private BigDecimal sortOrder;
     private List<RecommendSiteRespDTO> sites;
 }
